+++ common/src/main/java/io/github/mortuusars/envelope/world/mail/delivery/Courier.java
@@ -21,7 +21,8 @@ public interface Courier {
           new BaseDropOffHandler(),
           new BlockDropOffHandler(),
           new PlayerDropOffHandler(),
-          new EntityDropOffHandler()
+          new EntityDropOffHandler(),
+          context -> MailDropOffResult.returned(context.getMail())
     );
 
     Optional<Delivery> getCurrentDelivery();